Adjusted the linearProblem label so it matches the solver label.
Deleted the conditional orthogonalization parameter setting.
String comparisons with Gmres seem buggy. I used std::regex (c++11)
and with ignore case turned on, it seems to match the string on my
desktop, yet fail to match on Shepard. This resulted in several
wasted runs where the orthoType was not ICGS.