Work for adding better reporting to Teuchos TimeMonitor.
This is a WIP, and currently, the descriptive statistic code can
reprodude the existing aggregate values. The next step is to figure out
how to report the statistics. E.g., use the median for ranking and
report the min/max/mean median accross nodes.  We could also report
stats for ranks and link these to hostnames.

A few notes:
Num Observations does not equal numCalls, because the number of
observations is not always increased simply because the timer was
called. This occurs if the timers are nested. It is also not obvious how
to merge timers, or if that is even meaningful in the first place.