Remove print of warning of overwrite of build-stats file (#7376)
Turns out this breaks Fortran builds because it seems that the build system
uses the Fortran compiler to produce some output that it puts in a file and
then is read in by another compile comamnd with the Fortran compiler.
Therefore, these build wrappers must be completely silent except for what the
child command puts out.